Extract `SegmentEntry::take_snapshot` method into separate `SnapshotEntry` trait
