+++ lib/segment/src/entry/mod.rs
@@ -1,2 +1,7 @@
 pub mod entry_point;
 pub mod partial_snapshot_entry;
+pub mod snapshot_entry;
+
+pub use entry_point::SegmentEntry;
+pub use partial_snapshot_entry::PartialSnapshotEntry;
+pub use snapshot_entry::SnapshotEntry;
