+++ lib/segment/tests/integration/segment_on_disk_snapshot.rs
@@ -8,6 +8,7 @@ use rstest::rstest;
 use segment::data_types::index::{IntegerIndexParams, KeywordIndexParams};
 use segment::data_types::vectors::{DEFAULT_VECTOR_NAME, only_default_vector};
 use segment::entry::entry_point::SegmentEntry;
+use segment::entry::snapshot_entry::SnapshotEntry as _;
 use segment::index::hnsw_index::num_rayon_threads;
 use segment::json_path::JsonPath;
 use segment::segment::Segment;
